requirements.txt
## The following requirements were added by pip freeze:
appdirs==1.4.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1
django-compressor-autoprefixer==0.1.0
django-orderable==4.0.5
enum34==1.1.6
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
rcssmin==1.0.6
rjsmin==1.0.12
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
